Deprecate SelfResolvingDependency methods on ProjectDependency
A ProjectDependency is a SelfResolvingDependency, but in practice it is not resolved that way by the dependency resolution engine.
In order to support the self-resolving methods, a ProjectDependency needs an actual Project instance (or an instance to the project registry)
This should not be necessary. A project dependency should be a simple reference to the identifier of a project and no more.